refactor: user/answer/domain/dto 패키지를 user/answer/dto로 이동
