+++ src/main/java/com/upstage/devup/user/answer/dto/UserAnswerSaveRequest.java
@@ -1,4 +1,4 @@
-package com.upstage.devup.user.answer.domain.dto;
+package com.upstage.devup.user.answer.dto;
 
 import lombok.*;
