+++ src/main/java/com/upstage/devup/user/answer/service/UserAnswerSaveService.java
@@ -5,8 +5,8 @@ import com.upstage.devup.global.exception.EntityNotFoundException;
 import com.upstage.devup.question.dto.QuestionDetailDto;
 import com.upstage.devup.global.entity.Question;
 import com.upstage.devup.question.service.QuestionService;
-import com.upstage.devup.user.answer.domain.dto.UserAnswerDetailDto;
-import com.upstage.devup.user.answer.domain.dto.UserAnswerSaveRequest;
+import com.upstage.devup.user.answer.dto.UserAnswerDetailDto;
+import com.upstage.devup.user.answer.dto.UserAnswerSaveRequest;
 import com.upstage.devup.global.entity.UserAnswer;
 import com.upstage.devup.global.entity.UserAnswerStat;
 import com.upstage.devup.global.entity.UserWrongAnswer;
