+++ src/test/java/com/upstage/devup/user/answer/service/UserAnswerSaveServiceTest.java
@@ -3,8 +3,8 @@ package com.upstage.devup.user.answer.service;
 import com.upstage.devup.global.exception.EntityNotFoundException;
 import com.upstage.devup.question.dto.QuestionDetailDto;
 import com.upstage.devup.question.service.QuestionService;
-import com.upstage.devup.user.answer.domain.dto.UserAnswerDetailDto;
-import com.upstage.devup.user.answer.domain.dto.UserAnswerSaveRequest;
+import com.upstage.devup.user.answer.dto.UserAnswerDetailDto;
+import com.upstage.devup.user.answer.dto.UserAnswerSaveRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
